e2e: split xcm_transfers to different files
