+++ runtime/integration-tests/src/xcm/kusama/tests/mod.rs
@@ -1,0 +1,3 @@
+mod asset_registry;
+mod currency_id_convert;
+mod transfers;
