+++ runtime/integration-tests/src/xcm/kusama/tests/transfers.rs
@@ -412,213 +412,11 @@ fn transfer_foreign_sibling_to_altair() {
 	});
 }
 
-pub mod asset_registry {
-	use super::{assert_ok, parachains, GeneralKey, MultiLocation, X1};
-	use crate::xcm::kusama::test_net::{Altair, TestNet};
-	use altair_runtime::{Balance, CurrencyId, CustomMetadata, Origin, OrmlAssetRegistry};
-	use frame_support::assert_noop;
-	use orml_traits::asset_registry::AssetMetadata;
-	use orml_traits::currency::MultiCurrency;
-	use sp_runtime::traits::BadOrigin;
-	use xcm::prelude::{Parachain, X2};
-	use xcm::VersionedMultiLocation;
-	use xcm_emulator::TestExt;
-
-	#[test]
-	fn register_air_works() {
-		Altair::execute_with(|| {
-			let meta: AssetMetadata<Balance, CustomMetadata> = AssetMetadata {
-				decimals: 18,
-				name: "Altair".into(),
-				symbol: "AIR".into(),
-				existential_deposit: 1_000_000_000_000,
-				location: Some(VersionedMultiLocation::V1(MultiLocation::new(
-					0,
-					X1(GeneralKey(parachains::kusama::altair::AIR_KEY.to_vec())),
-				))),
-				additional: CustomMetadata::default(),
-			};
-
-			assert_ok!(OrmlAssetRegistry::register_asset(
-				Origin::root(),
-				meta,
-				Some(CurrencyId::Native)
-			));
-		});
-	}
-
-	#[test]
-	fn register_foreign_asset_works() {
-		Altair::execute_with(|| {
-			let meta: AssetMetadata<Balance, CustomMetadata> = AssetMetadata {
-				decimals: 12,
-				name: "Acala Dollar".into(),
-				symbol: "AUSD".into(),
-				existential_deposit: 1_000_000_000_000,
-				location: Some(VersionedMultiLocation::V1(MultiLocation::new(
-					1,
-					X2(
-						Parachain(2000),
-						GeneralKey(parachains::kusama::altair::AIR_KEY.to_vec()),
-					),
-				))),
-				additional: CustomMetadata::default(),
-			};
-
-			assert_ok!(OrmlAssetRegistry::register_asset(
-				Origin::root(),
-				meta,
-				Some(CurrencyId::ForeignAsset(42))
-			));
-		});
-	}
-
-	#[test]
-	// Verify that registering tranche tokens is not allowed through extrinsics
-	fn register_tranche_asset_blocked() {
-		Altair::execute_with(|| {
-			let meta: AssetMetadata<Balance, CustomMetadata> = AssetMetadata {
-				decimals: 12,
-				name: "Tranche Token 1".into(),
-				symbol: "TRNCH".into(),
-				existential_deposit: 1_000_000_000_000,
-				location: Some(VersionedMultiLocation::V1(MultiLocation::new(
-					1,
-					X2(Parachain(2000), GeneralKey(vec![42])),
-				))),
-				additional: CustomMetadata::default(),
-			};
-
-			// It fails with `BadOrigin` even when submitted with `Origin::root` since we only
-			// allow for tranche tokens to be registered through the pools pallet.
-			let asset_id = CurrencyId::Tranche(42, [42u8; 16]);
-			assert_noop!(
-				OrmlAssetRegistry::register_asset(
-					Origin::root(),
-					meta.clone(),
-					Some(asset_id.clone())
-				),
-				BadOrigin
-			);
-		});
-	}
-}
-
 #[test]
 fn test_total_fee() {
 	assert_eq!(air_fee(), 9324000000000000);
 	assert_eq!(fee(decimals::AUSD), 9324000000);
 	assert_eq!(fee(decimals::KSM), 9324000000);
-}
-
-pub mod currency_id_convert {
-	use super::*;
-	use altair_runtime::CurrencyIdConvert;
-	use sp_runtime::traits::Convert as C2;
-	use xcm::VersionedMultiLocation;
-	use xcm_executor::traits::Convert as C1;
-
-	#[test]
-	fn convert_air() {
-		assert_eq!(parachains::kusama::altair::AIR_KEY.to_vec(), vec![0, 1]);
-
-		// The way AIR is represented relative within the Altair runtime
-		let air_location_inner: MultiLocation = MultiLocation::new(
-			0,
-			X1(GeneralKey(parachains::kusama::altair::AIR_KEY.to_vec())),
-		);
-
-		assert_eq!(
-			<CurrencyIdConvert as C1<_, _>>::convert(air_location_inner),
-			Ok(CurrencyId::Native),
-		);
-
-		// The canonical way AIR is represented out in the wild
-		let air_location_canonical: MultiLocation = MultiLocation::new(
-			1,
-			X2(
-				Parachain(parachains::kusama::altair::ID),
-				GeneralKey(parachains::kusama::altair::AIR_KEY.to_vec()),
-			),
-		);
-
-		Altair::execute_with(|| {
-			assert_eq!(
-				<CurrencyIdConvert as C2<_, _>>::convert(CurrencyId::Native),
-				Some(air_location_canonical)
-			)
-		});
-	}
-
-	#[test]
-	fn convert_ausd() {
-		assert_eq!(parachains::kusama::karura::AUSD_KEY.to_vec(), vec![0, 129]);
-
-		let ausd_location: MultiLocation = MultiLocation::new(
-			1,
-			X2(
-				Parachain(parachains::kusama::karura::ID),
-				GeneralKey(parachains::kusama::karura::AUSD_KEY.to_vec()),
-			),
-		);
-
-		Altair::execute_with(|| {
-			assert_eq!(
-				<CurrencyIdConvert as C1<_, _>>::convert(ausd_location.clone()),
-				Ok(CurrencyId::AUSD),
-			);
-
-			assert_eq!(
-				<CurrencyIdConvert as C2<_, _>>::convert(CurrencyId::AUSD),
-				Some(ausd_location)
-			);
-		});
-	}
-
-	#[test]
-	fn convert_ksm() {
-		let ksm_location: MultiLocation = MultiLocation::parent().into();
-
-		Altair::execute_with(|| {
-			assert_eq!(
-				<CurrencyIdConvert as C1<_, _>>::convert(ksm_location.clone()),
-				Ok(CurrencyId::KSM),
-			);
-
-			assert_eq!(
-				<CurrencyIdConvert as C2<_, _>>::convert(CurrencyId::KSM),
-				Some(ksm_location)
-			)
-		});
-	}
-
-	#[test]
-	fn convert_unkown_multilocation() {
-		let unknown_location: MultiLocation = MultiLocation::new(
-			1,
-			X2(
-				Parachain(parachains::kusama::altair::ID),
-				GeneralKey([42].to_vec()),
-			),
-		);
-
-		Altair::execute_with(|| {
-			assert!(<CurrencyIdConvert as C1<_, _>>::convert(unknown_location.clone()).is_err());
-		});
-	}
-
-	#[test]
-	fn convert_unsupported_currency() {
-		Altair::execute_with(|| {
-			assert_eq!(
-				<CurrencyIdConvert as C2<_, _>>::convert(CurrencyId::Tranche(
-					0,
-					[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]
-				)),
-				None
-			)
-		});
-	}
 }
 
 fn air_fee() -> Balance {
@@ -640,7 +438,7 @@ fn ksm_fee() -> Balance {
 
 fn calc_fee(fee_per_second: Balance) -> Balance {
 	// We divide the fee to align its unit and multiply by 4 as that seems to be the unit of
-	// time the transfers take.
+	// time the tests take.
 	// NOTE: it is possible that in different machines this value may differ. We shall see.
 	fee_per_second.div_euclid(10_000) * 8
 }
